dev mode that makes it possible to use passport locally with no integrations (#30)
This pull request does two things:
1) documents a bunch of stuff in readmes
2) introduces a new environment variable named
`BYPASS_EMAIL_REGISTRATION` which enables local development without
connecting to either pretix or sendgrid. basically it allows any email,
even an email that isn't a valid email, to log into the passport and use
it locally. the reason i think this is necessary is so that we can point
3rd party developers to the repo, and have them build and test against a
local version of the passport server and client, so they don't have to
be blocked by whatever we are doing at zuzalu.