generic prove screen (#89)
#### 🚨we'll need to publish new versions of packages after deploying
this🚨
#### summary
The primary purpose of this PR is to make generic some of the
functionality of this repository that was deliberately made specifically
for Zuzalu. Additionally, I create better documentation, fix some
hard-coded things, and extract some code for reuse. Here's a more
complete list of changes:
- created a `ProofOptions` interface, by which third party applications
can configure how their PCD request is communicated to the user. One
important field of this interface is `genericProveScreen`. if the 3rd
party developer sets this to `true`, their PCD request is routed to the
`GenericProveScreen`, which is designed to be able to receive user input
for any type of PCD, current or future.
- one small thing I'm working on here is a `debug` field, which is not
super fleshed out rn, but I hope will be useful during the hackathons
and such
- fleshed out `ParameterizedProofScreen`, and renamed it to
`GenericProveScreen`. This screen can handle any type of PCD, and I've
started working on error handling for it, so that it can validate user
input. It also handles disabling user input if args values are provided
by the third party application, and it handles loading values from
remote locations, which is used for the semaphore group, for example.
- I updated the consumer client application to have more of a
'directory' structure. it has two sections now: one section for PCD
requests that go through the Zuzalu-specific prove pages, and one for
'generic' proofs, which are basically the same except they go through
the `GenericProofScreen`.
- I also updated the consumer-client application to look a little more
clean - JSON responses are now collapsible, examples include more
English description of what's going on.
- I factored out the code that receives `message`s from the proof popup
and turned it into a hook that can be imported by third party
applications.
- The generic proof screen is not pretty, and we should probably get Jon
to design something better, that takes into account the features and
constraints we want (input validation, remote loading, debug mode,
hiding/showing certain inputs, etc.). Here's a screenshot:

<img width="493" alt="Screenshot 2023-03-30 at 3 52 48 PM"
src="https://user-images.githubusercontent.com/2636237/228783491-66a26fe7-5918-456c-85a8-062d1508038e.png">

#### video demo

https://user-images.githubusercontent.com/2636237/228779884-6de5cd42-54ea-4aa4-b903-d7ef1b4b7cc2.mov