fix npe on passport reset (#237)
### summary

This PR fixes two issues that I observed in rollbar. Neither of them
causes user-visible errors.


#### issue number one
rollbar links: 
- https://app.rollbar.com/a/0xparc/fix/item/zupass/52
- https://app.rollbar.com/a/0xparc/fix/item/zupass/53

in some cases, the user does not have a sync key saved in local storage.
i have only been able to reproduce this by clearing the passport using
the 'clear passport' button, but I suspect there are some other cases
where there is no sync key.

the fix for this issue is the first commit of this pr

#### issue number two
rollbar links (i'm not sure how to merge issues in rollbar)
- https://app.rollbar.com/a/0xparc/fix/item/zupass/54
- https://app.rollbar.com/a/0xparc/fix/item/zupass/55
- https://app.rollbar.com/a/0xparc/fix/item/zupass/56
- https://app.rollbar.com/a/0xparc/fix/item/zupass/57


in some cases, the user has a valid passport without having uploaded
their passport data to e2ee for some reason. i think I made it less
likely for a user to end up in this state in my previous sync pr by
making the passport to save e2ee before redirecting the user to the home
page with all their pcds only after their pcds have been uploaded to
e2ee.

however, i think it is still possible for someone to end up in this
state if they refresh the page after signing up, but before the upload
to e2ee has completed.

so, that takes me to commit number two, which ensures that the sync goes
through even in the case that there hasn't been a valid e2ee upload
before, by catching the exception, and setting some local state that
allows the sync to proceed.

I was able to reproduce this issue locally by deleting e2ee storage from
the database. the error only appeared in the console, and did not break
user-facing functionality, except for sync.