display modals as full-screen overlays in the popup screens (#258)
## Summary

closes https://github.com/proofcarryingdata/zupass/issues/255

If a user resets their passport by signing into another device, the
passport they originally signed in on is 'invalidated'. `zupass.org`
asks 'invalidated' passports to sign out, when you go to the home
screen, but it did not ask users to sign out when they open up a login
popup. This caused a bunch of confusion - people attempt to sign in and
the passport lets them, even though they are not signed in properly to
zupass.org.

This PR fixes this issue by displaying the 'invalid state' popup as
full-screen overlay modals in the popups.

<img width="765" alt="Screenshot 2023-05-03 at 2 20 05 PM"
src="https://user-images.githubusercontent.com/2636237/235913856-0605ca1d-b34f-4a08-8e35-550194fd84fd.png">