Stricter sync checks for missing events/items (#451)
Closes #317 

Sync would already fail for missing events, because we iterate through
all configured events and fetch them directly from the server, throwing
an exception of the request fails/event is not found.

Sync now fails if an event has an `activeItemID` or `superuserItemID`
that refers to an item which does not exist in Pretix. This error is
detected during validation of fetched data, by comparing the fetched
item IDs to the configured ones.

This PR also adds tests to verify that changes to event and product
names are reflected in ticket PCDs.