Fix all lints (#461)
Fixes #428 

The majority of these are non-null assertions. These break down mostly
into two categories: one, where we're `.get()`ing something from a map,
often as part of a test, and we know that the item is present even
though the type-checker doesn't. In this cases, `.get("id") as TYPE` has
the same effect but without causing a linting warning.

The second is `init()` calls on PCD packages. The problem is that the
base PCD package interface defines its function members as optional, and
even when the concrete class defines them, the type-checker still thinks
it's optional. The non-null assertion gets around the type-checker, but
our ESlint rules don't like non-null assertions. Converting to optional
invocations (`?.()`) satisfies the type-checker. However, this will
produce slightly different JS output, where if the function is in fact
undefined, no attempt will be made to call it. This would mean that no
run-time error would occur.

In practice, this seems very unlikely to happen, because we're only
calling `init()` for packages where we know that we've defined it.
However, we could replace it with a run-time assertion (`if
(!somePackage.init) throw new Error(...)`) which would also satisfy the
type-checker and cause a run-time error, at the cost of making each
invocation of `init` a bit noisier. I'd be happy to switch it to that
approach.