Simplify handling of 429 error and cancellation of pending requests (#488)
Fixes #467 and #468

Implementing a retry for 429 errors seems unnecessary, and is hard to
test. Instead, we can log that it happened and perhaps adjust the limits
on client-side rate-limiting to ensure that it doesn't happen again.

Separately, it seems that `fetch` does not remove event listeners in a
timely fashion, so if we make a large number of sequential calls to
`fetch`, we can easily end up with > 10 event listeners on a single
`AbortController` signal, which causes a warning to be logged. In this
implementation, each `fetch` gets its own `AbortController`, with a
separate `AbortController` which is only used when cancelling the
pending requests, and for which we manage event listeners manually.