Add `/telegram/message` endpoint for anonymous chat app (#486)
This PR adds a `/telegram/message` endpoint, which should be hit after
generating a ZKEdDSA PCD. The handler verifies the PCD, checks the
event, then sends the anonymous message to the "q & a" channel.

Main TODOs left are
- [ ] Frontend - I have something simple (see attached image below), and
wondering the best way to add this. Whether it's a simple HTML / React
page on a Passport route, or just a totally separate web app I can host.
@ichub let me know thoughts
- [ ] Programmatically sending messages in channel is hardcoded - see
the TODO in the `sendToAnonymousChannel` function. If there isn't a
better approach, we can proceed with adding a column to the events
table.

<img width="762" alt="Screenshot 2023-08-31 at 1 08 47 AM"
src="https://github.com/proofcarryingdata/zupass/assets/36896271/05873725-1c56-4f1b-b184-68b78c558dcf">

---------

Co-authored-by: Ivan Chub <ivansergchub@gmail.com>