sql-backed cache + cache issued tickets (#494)
it takes 5-6 seconds for the server to issue me tickets on pcdpass.xyz

this pr caches tickets so that they are generated only once. the cache
uniquely identifies the tickets by the hash of their `JSON.stringify`'d
fields, excluding the `timestampSigned` field

on my computer, this pr doubles the speed of issuance on the happy path
(which should be most of the paths)

in production I expect this will make issuance go down more
significantly.