Aggregate logging of related errors (#612)
Closes #555

This simplifies logging during sync. Previously, we were throwing
exceptions, catching them, logging the exception, then returning a
boolean `false` to a higher level function, which would then throw an
exception, that would get logged, and so on.

Because these exceptions get logged separately, Rollbar doesn't know
that they're connected. As a result, the higher-level log messages are
pretty useless, saying that a failure occurred but with no context as to
why.

This PR removes all of the lower-level logging to Rollbar. Instead,
exceptions bubble up and get logged by a high-level handler - in this
case, the `syncSingleOrganizer` method of `DevconnectPretixSyncService`.
If intermediate functions need to add context, then they can catch and
throw a new exception that gives the original as the `cause`.

In the `reportError` method of `RollbarService`, we take the causal
chain of exceptions and pass them all to a single call to
`Rollbar.error`, which means that Rollbar knows how to display them as a
single error in the UI:

<img width="587" alt="image"
src="https://github.com/proofcarryingdata/zupass/assets/20022/a092db90-080c-41b8-a97c-226de4427312">

The outcome here is that all sync errors will appear under a single
heading "Devconnect Pretix Sync failed", and each individual occurrence
might have a different stack trace containing the details of what caused
the error.