Stop the correct sync service during testing (#604)
Playing around with [Knip](https://github.com/webpro/knip) to see if I
could find any unused code, I noticed that it said that
`DevconnectPretixSyncService.stop()` is unused. "That's weird", I
thought, "because we definitely stop the service in
`devconnect.spec.ts`". It turns out that we don't! We're actually
stopping `pretixService`, not `devconnectPretixService`. This PR fixes
that.