Zuzalu tickets in PCDPass (#644)
Closes #618

This PR adopts a similar approach to verified email. There is a new
feed, which serves up Zuzalu ticket PCDs, based on the EdDSATicketPCD.

For this to work, we have to enable the Pretix service (distinct from
Devconnect Pretix service) in PCDPass. Local environment variables
(`PRETIX_*`) will need to be set with working values as per Zupass.

Out of scope for this PR is anything that allows the user to sign in to
Zuzalu.City or perform Zuzalu-specific actions with their ticket - this
just gives them a ticket in their wallet.

The feed has not been added as a "default subscription", so subscribing
requires manually subscribing (in the same manner as the Frog feed).

All Zuzalu use cases have been tested in `consumer-client`.