fix flickering background color (#603)
When loading PCDPass app, we always get the `--bg-dark-primary` bg color
from the static css file first. However, many screens use `gray` bg
colors and it creates a odd experience when loading a path directly. For
example, opening the /prove link from Telegram, one would see a
flickering background color.

If we simply remove the background color in the stylesheet, the
background color ended up to be a flash of white on first paint because
AppContainer is not yet loaded. Indeed, this is a common problem
especially for websites with dark mode:
https://stackoverflow.com/questions/37126401/modify-bodys-classes-before-first-paint

This PR introduces an eagerly executed js blob that pattern matches on
location hash. This ensure the bg color is correctly set on first paint.
It also refactor the AppContainer bg to switch on a centralized list of
routes config. We still want to keep the AppContainer injection because
the hand written js might not always give the correct result as the
HashRouter.

Tested locally with both `yarn dev` and `yarn build` to ensure build can
access the config file.

See the comparison between this branch and production build



https://github.com/proofcarryingdata/zupass/assets/4317392/6d952d0b-27a7-4b64-a53f-2029f19c187a

relates to https://github.com/proofcarryingdata/zupass/issues/597

---------

Co-authored-by: Forest Fang <4317392+saurfang@users.noreply.github.com>