Detect missing or unused dependencies, unused exports, files etc. (#609)
This PR comes from noticing an unused function in another PR, and
thinking about how we could detect unused exports. Non-exported unused
functions are detected by ESLint, but if a function is exported then
ESLint won't complain even if that export isn't used anywhere.

[Knip](https://github.com/webpro/knip) is a tool which detects those
kinds of issues across a whole codebase, including monorepos. It can
detect:
- Unused exports and types
- Unused files
- Unused dependencies and devDependencies
- Missing dependencies and devDependencies
- And some other less important stuff

The missing dependencies part is a particular issue for monorepos,
because we can have a package which uses code from a dependency (either
another package in the monorepo, or a third-party dependency) without
declaring it in `package.json`. If each package had its own repo, this
problem would become apparent quickly, but in monorepo setups the code
will still build even if the dependency is not declared, so long as some
other package does declare the dependency (or the code being depended on
is local to the monorepo). This means that our apps will build, but
individual NPM packages we create will not list their dependencies
correctly.

This PR is broken down into 5 commits:

1. Adds a `yarn knip` script to package.json and a basic `knip.json`
config file. Checking out this commit lets you run `knip` against our
codebase without any changes.
2. Adds in the missing dependency declarations, so each package properly
declares what it depends on
3. Removes unused dependencies, so where a package declares a dependency
on some package but never actually uses it, that dependency is removed
4. Same as 3 but for devDependencies (with some ESLint dependencies left
even though they flag up as unused; I think this part requires some
interpretation).
5. Adds `yarn knip` to the CI process so that we can't reintroduce the
problems mentioned above

In the final config, missing dependencies and unnecessary dependencies
are flagged as errors that would block CI. All other issues are flagged
as warnings. We might want to move to flagging unused exports or types
as errors, but that seems excessive.

At this point, a number of exports are flagged as unused. Specifically,
this means that they are unused _outside of the file they are declared
in_. So if I declare a function as exported, but use that function
within the same file, it will be flagged as an unused export because the
`export` declaration is unnecessary. In some cases, the function will be
entirely unused, in which case removing the `export` declaration will
allow ESLint to detect it as an unused function.