Redirect users to login if accessing subscription screens while logged out (#660)
Closes #537 and #566 

When accessing the subscription screens in passport-client while logged
out, the user is now redirected to login (or create an account). Upon
logging in (or creating an account) the user is then redirected back to
the subscription screen that they attempted to access.

This will allow deep links to the subscription screen to be provided,
e.g. for instructing Zuconnect attendees on how to add the Zuconnect
ticket feed.