Fix flaky Ethereum ownership PCD test (#675)
Fixes #668

The test wants to cause verification to fail, and does so by tampering
with the Ethereum address. The bug is that it does this by replacing the
final character of the address with the constant "0", which works...
except when the final character of the address is already "0", hence the
intermittency of the failure. This checks for this case, and uses a "1"
instead.