non-login/account-related pcdpass->zupass transition (#710)
#### summary
progress on #699 - by the end of the week I aim to fully remove the
`isZuzalu` checks in the backend and frontend. rather than doing it all
in one go I'm going to break it up into small and easily digestible
changes.

these changes update the frontend and backend of zupass to no longer
make a distinction between zupass and pcdpass in the cases where that
change was simple.

i have stopped here, because the rest of the changes relate to login and
account management, which is something that is being actively worked on
by @rrrliu, whom I don't want to interfere with too much.

these changes are relatively rote, and don't really require much
thought. I've annotated this PR to highlight the changes that ppl need
to be aware of

#### testing
this pr includes some changes to tests. my intention is to merge the
`devconnect.spec.ts` and `zupass.spec.ts` tests after the
pcdpass<->zupass merge is complete. until then, i've updated the the
zupass and pcdpass tests to run both pretix sync services, and so that
both request issued pcds via the feeds api, and both are able to serve
the old zuzalu tickets