prevent NPE in sync code (#822)
https://github.com/proofcarryingdata/zupass/issues/642 would have caught
this. I think we should do it before zuconnect.

the way this bug manifested for me is that when I logged in with a new
email, or reset an existing account, the app would not finish going
through the sync process, because we had never saved the e2ee storage
for the user. refreshing would then load it properly.