Very simple caching of verifications (#820)
Closes #813 

I spent a while thinking about how to do bulk feed fetching, and then
came to the conclusion that caching verifications is much better. For
instance, error handling gets complicated when there are multiple feeds,
and clients need to know how to aggregate all requests which share a
credential - neither are fiendishly complicated but they add a lot of
incidental complexity to what are otherwise quite clean and simple
systems.

This is the simplest thing that could possibly work, but it does seem
highly effective. Testing 100 simultaneous requests with `yarn scratch`
_without_ caching (10 workers, on an M1 Macbook Air):

`✨  Done in 43.40s.`

_With_ caching:

`✨  Done in 2.41s.`

Now, 100 simultaneous requests for verification of the same PCD is a
rare scenario. But three or four simultaneous requests is quite common,
if someone is subscribed to all of the currently-available Zupass feeds
then they will do this every time they refresh their feeds. Because the
cache caches _promises_, we don't have to wait for the first
verification to succeed before we can store it in the cache. On the
first request, we create a promise and cache it, and the subsequent
requests get served the same promise, without tying up any more than the
original worker process. When it resolves, they all get the same
verification result. They can even turn up later and request it again,
and the promise will continue to resolve for as long as it remains in
the cache.

Todo: tests of the LRU cache. Is testing the multiprocess functionality
easy or are there pitfalls there?