[DMS/ZMS] Link multiple anon topics (#814)
Creates a new table, `telegram_chat_anon_topics`, where the primary key
is `(ticket_event_id, anon_topic_id)`. We can't use `telegram_chat_id`
in `telegram_bot_events` as a foreign key because it's not unique.

Make sure to set `TELEGRAM_ANON_BOT_WEBAPP` in `.env` for
`passport-server`. If you don't know where to get this link, ask
@cha0sg0d or @lermchair.

This does not pre-fill out the ticket in Zupass, which is blocked by
#762

---------

Co-authored-by: cha0sg0d <0xcha0sg0d@gmail.com>
Co-authored-by: cha0sg0d <42984734+cha0sg0d@users.noreply.github.com>