Cache temporary credentials in local storage (#896)
Closes #639 

This eliminates some instances of expensive `prove` operations during
startup.

If the user has generated a feed credential within the last hour, there
is no reason for them to generate another one. By persisting the cache
to local storage and repopulating the cache from local storage on
initialization, we can reuse credentials generated on the previous page
load or in another tab.