Safer service worker (#976)
The primary feature here is that the service worker deletes all cache
entries which it doesn't explicitly recognize at install time. This will
avoid future instances of the problem @robknight encountered with a
stale cached javascript bundle.

I tested the normal-case behavior by observing the new log statements in
the console. I tested the dangerous case by temporarily uncommenting the
line to cause index.js to be added to the cache (visible in devtools),
then commenting it out again, and watching it be discarded on next
reload (confirmed in devtools).

Secondarily, I fleshed out the list of intentionally-cached files with
more files I've observed in testing, including the artifacts from the
ZKEdDSAEventTicket PCD.