ZK QR codes (#951)
Closes #772 
Closes #943 

Zuzalu QR codes are now ZK by default. Zuconnect and Devconnect now use
simple QR codes by default, with "ZK mode" using a URL that passes
serialized `ZKEdDSAEventTicketPCD` to passport-client.

Scanning either the simple or ZK versions of a Zuconnect ticket, or a ZK
version of a Zuzalu ticket, takes you to a "verify" screen which will
either receive the ticketId in the query, or extract it from the PCD's
revealed fields. TicketIds in the query are stored alongside a timestamp
in a stringified JSON object which is then Base64 encoded, to prevent
trivial re-use of the QR code.

Scanning either the simple or ZK versions of a Devconnect ticket takes
you to a check-in screen which will either receive the ticketId in the
query, or extract it from the PCD's revealed fields. Check-in works
identically irrespective of how the ticketId is communicated.

---------

Co-authored-by: Ivan Chub <ichub@users.noreply.github.com>