Fix flaky sync test (#1003)
I'm not quite sure why this has stopped working now, but it seems to be
broken across all branches. The old code works locally, so perhaps the
failure is due to something that has changed in the Github CI
environment.

If we merge this, I will create a new issue for further investigation.