Service worker for offline viewing (#980)
Experimental service worker functionality for offline PCD viewing. This
isn't complete, but it works for all basic needs, incorporates new
discoveries from today's prod rollout, and has TODOs which properly
reflect the pieces I want to address before I consider it ready to
merge. I think it's worth getting a first round of feedback from code
review while I keep working on those items.

Caching behavior is as follows:
- For items defined to be "stable", cache hits are returned immediately,
skipping the network. These update only when a new worker is installed.
This minimizes bandwidth and delays for large static items like wasm.
- For items defined to be "ephemeral", network is tried first, with
cache as fallback on failure. These are updated after any successful
network fetch. This ensures timely updates for page contents and code.
- Other items (limited to GET requests within the domain) will be added
to the ephemeral cache as they are encountered. This is optional if our
list of ephemeral items is complete, but seems a reasonable best-effort
in case we miss anything.

This is enough for the app to function well offline to view PCDs.
Adding/editing PCDs works offline too, including proving. However any
modifications will be blown away by the download on the next
online-fetch. That can be solved separately by the use of the revision
number in E2EE sync, but isn't a goal to fix in this PR.

I don't like how big and complex this code has gotten, but I think it's
necessary for all the feature decisions so far. Suggestions are welcome.