New screen to display a nice error when devices don't support WASM (#1017)
Closes #956

This pull request adds a check in the Zupass app to verify if the device
supports WebAssembly (WASM). In the event that WASM is not supported, a
friendly error message will be displayed to the user.

Without WASM, Zupass cannot function effectively, as
`@pcd/passport-crypto` (specifically `libsodium-wrappers-sumo`) relies
on WASM and is used in the login process. Additionally, all PCD packages
that make use of `snarkjs` and `circomlibjs`, which is almost all of
them, would not work.

The screenshot below displays the message that appears when the device
does not support WASM.

![Screenshot from 2023-10-19
16-30-56](https://github.com/proofcarryingdata/zupass/assets/11427903/a60be403-5043-41c0-967f-e93e0e559734)

---------

Co-authored-by: Richard Liu <richard.y.liu@berkeley.edu>