Fix flash (#1061)
Closes #833 

The update to `self` is asynchronous, so there might be a small delay
before the redirection to the home screen takes place, especially if
you're on a slow network. To provide feedback to the user, we need to
keep displaying a loader instead of clearing it early.

### How to test:

In your dev tools (on any branch that's not this one), throttle your
network to something slow (e.g. Slow 3G) and then log in. You should see
a brief "flash" of the login screen after entering your password and
clicking the login button.

Try again but on this branch. There should be no flash.