Redacted tickets (#988)
Closes #910 

When logging in as a new user, you are prompted to agree to the legal
terms before (optionally) setting a password.

When logging in as an existing user, if you have not previously agreed
to the terms, you are prompted to do so with an un-dismissable modal.
This works in a similar manner to the `RequireAddPasswordModal`, by
inspecting the user object for valid values and displaying a modal if
invalid values are found.

Agreeing to the terms as an existing user causes an API request to the
backend. If this request fails, the terms agreed are persisted to local
storage, and the API request will be tried again on next load.

Devconnect pretix sync now keeps a list of the users who have agreed to
the terms, and during ticket sync will redact any tickets where the
attendee's email is not in this list. In the case where the attendee has
not agreed to the terms, their ticket is stored in a separate table,
with a hashed email and no name.

On first-time login, or on the API request to agree to the terms,
tickets are un-redacted by copying data from the redacted ticket table
to the full ticket table. On the client-side, agreement to the terms
causes a refresh of feeds to be performed, meaning that un-redacted
tickets should appear within a second or two (depending mostly on
latency).

Ticket redaction is enabled by an environment variable. However, this
only affects the redaction and does not affect the legal terms agreement
interface. There ought to be no visible difference to the user in either
case.

---------

Co-authored-by: Ivan Chub <ivansergchub@gmail.com>