Fix accidental detection of differences between Zuconnect tickets (#1091)
When saving a Zuconnect ticket to the DB, we give it a default
`extra_info` value of an empty array. However, the corresponding API
ticket has value of `undefined`, and the comparison function did not
account for this. Now we assign an empty array to this value during
parsing, and make the comparison function do a deep equality check.