Remove hard-coded QR code width (#1102)
In in-person testing, I've seen a case of someone whose phone resolution
is low enough that they can't display a 300px-wide QR code.

The 300px width being hard-coded seems to be unnecessary, and removing
it doesn't cause a problem on other browsers, but does allow it to be
scaled down on screens with fewer than 300px available.