ZuAuth: new authentication example in consumer app (#1079)
Closes #911, #1092

The video below demonstrates how a single click can allow users to
authenticate using Zupass tickets anonymously with a ZK proof
(`@pcd/zk-eddsa-event-ticket-pcd`) or with a simple Semaphore signature
(`@pcd/semaphore-signature-pcd`). The consumer server validates the PCD
and establishes a session. The session remains active upon refreshing
and is terminated upon logging out, as you can see on the right in the
cookies section.

[Screencast from 24-10-2023
12:57:43.webm](https://github.com/proofcarryingdata/zupass/assets/11427903/9d358223-5a52-4579-87d5-73391eb4bddc)

The sequence diagram below illustrates how the components interact with
each other. In summary:
1. The consumer client requests the consumer server to generate a
session nonce and to include it within the PCD (the nonce is the
watermark and the external nullified in zk mode).
2. The consumer client opens the Zupass popup to create the PCD.
3. The newly generated PCD is sent to the consumer server to complete
the login process.
4. The consumer server validates the PCD and its parameters and
generates a user session cookie.
5. The user now has a valid session and can access features reserved for
logged-in users.

```mermaid
sequenceDiagram
    Consumer-client->>Consumer-server: getNonce()
    Consumer-server->>Consumer-client: return <<nonce>>
    Consumer-client->>Zupass: requestPCD()
    Zupass->>Consumer-client: return <<PCD>>
    Consumer-client->>Consumer-server: login()
    Consumer-server->>Consumer-client: return <<cookie>>
```

---------

Co-authored-by: Jeeiii <giacomo.corrias7@gmail.com>