More stable uploadID to avoid extra uploads. (#1139)
This fixes some of the causes behind #1073, so we should be able to do
fewer unnecessary uploads, meaning fewer opportunities to clobber the
user's changes.

The primary fix here is to the Email PCD. Each time it was re-issued
through feeds, it had different contents (specifically the random UUID
in the embedded EdDSA signature PCD), thus causing the user's PCDs to be
different and need to be uploaded.

There's also a minor fix to save the uploadedID after the initial upload
on account creation, which will keep the PCDs from needing to be
uploaded twice for that case. I also deleted an unused bit of state
tracking which was relying on the old uploadID format.

The rest of the changes here are enhancements to tests which I wrote
during the process of tracking down the cause of the changes. More test
coverage is probably good, so I'm keeping them.