Delete leftover redacted tickets (#1152)
In cases where a user unredacts their tickets during sync, new redacted
tickets can be created for them. These are harmless and don't interfere
with their unredacted tickets, but there's no good reason for them to be
there.

Sync _ought_ to clear up redacted tickets that serve no purpose, but it
turns out that `node-postgres` doesn't do what one might expect with
`IN($1)` in queries where `$1` is an array, and `= ANY($1)` [is the
preferred
syntax](https://github.com/brianc/node-postgres/wiki/FAQ#11-how-do-i-build-a-where-foo-in--query-to-find-rows-matching-an-array-of-values).
This was causing the deletion of redundant redacted tickets to fail
silently.

A new test has been added to catch this.