Track server-storage-hash persistently to avoid extra uploads (#1167)
Introduces server-storage-hash as replacement for uploadId. It is
calculated directly from the serialized (unencrypted) blob, so should be
more certain to match what's changed, and not miss things (such as the
user object). It's stored in AppState and local storage, beside
server-storage-revision. I added an explicitly-stable JSON serializer
for the top-level structures to help ensure the hash stays consistent,
though note that I haven't applied it in the PCDs themselves, so we're
still relying on them not being dynamically updated in a way which would
change serialization order.

With these changes it's finally to go through reloads, logoff, and login
without ever doing an extra upload/download, and thus without risking
clobbering changes. The revision number remains unchanged through those
operations. Password change does one revision change (to change the blob
key) but no extra ones. Closes #1073.

I'm concerned that the tracking of sync status in AppState leads to
delayed updates which can be unpredictable. The final hack I used to fix
the password change flow (via using update() rather than dispatch() to
perform updates) was to fix an issue of that type. I think there's a
cleaner long-term design, but I don't want to upset the overall design
too much at this point in prod rollout.