Service Worker timeout should cover full download time (#1202)
2 functional changes:
- Wait for cache update to complete inside of startFetch(). This ensures
download is completed inside the scope of the timeout.
- Look up in ephemeral cache before starting fetch. This avoids a
theoretical risk of collision between reading the cache on timeout and
updating the cache in the background.

Also added zxcvbn.js to the list of stable resources, so we only ever
have to fetch it once.

With these changes, and throttling enabled in chrome devtools, I see the
cache timeout occur during the long (>1m) download of index.js. So
user-visible load time is about 7s, after the timeout of 5s. Note that I
saw inconsistency in whether or not Chrome devtools' throttling applied
to the service-worker, so be careful if you try to test that way, to be
sure it's really working.