Periodic fetch of feeds in background. (#1207)
Add an extra fetch of subscription feeds each time the periodic poll is
triggered (every 60s while the app is visible).

The mechanism within the client is identical to what's already done for
downloads from E2EE storage, so should be pretty sfae.. The difference
is that feeds don't have anything in the network protocol to avoid
re-downloading the same actions. However the use of hashes in the client
is now stable enough that applying the same actions repeatedly won't
make the PCD collection appear to be different, so won't trigger any
extra uploads which could clobber changes.

With that in mind, I also reduced the frequency of the legacy "Frogs"
feed to 24h so it doesn't fill a user's Zupass with frogs generated by
background polls. We may want to remove that feed entirely now that
Frogcrypto is live, but that's outside the scope of this PR.