Add ZK EdDSA Frog PCD (#1162)
Added a new ZK EdDSA Frog PCD representing a proof of ownership of an
EdDSA-signed frog.
The prover is able to prove ownership of a frog corresponding to their
semaphore identity, and keep their identity private.
The proof can also include a nullifier.

This will be used for telegram integration. For example, you can verify
you own a frog or a specific type of frog in order to join a telegram
channel, without revealing your semaphore identity.

TODO: need to add watermark for telegramUserId