[frogcrypto] Fix image zoom (#1254)
![image](https://github.com/proofcarryingdata/zupass/assets/10219618/5b818ceb-d220-4e52-9683-c346b7759bd8)

When you try to expand the image in the modal, this happens because the
modal has a z-index of 999, so making the z-index 1000 for the image
fixes that issue.

I was not able to run this because I couldn't install turbo, so please
check it first. I have no idea if this will work.