Remove hash from emitter in PCDCollection (#1288)
This feature is vestigial, no longer used. App updates on PCD changes
are driven by calls to update(). [Update: this isn't true, the change
callback is needed by React, but the hash is not. See further comments
below.]

Repeatedly serializing all PCDs was causing slowdowns, including in
applying feed actions. Closes #1270.

On my local dev server, on an account with 800+ frogs, this change takes
the `applyActions()` part of sync from ~1800ms to ~25ms.