Generic rate limiting (#1264)
Closes #1261

This implements a generic rate limiting feature, with rate limits
applied to both requests for, and verification of, email tokens (which
are used for initial email verification and for account reset).

I considered a few options here, including:
- Fixed window rate limiting, where we detect the first instance of a
rate-limited action, save the timestamp, and then increment a counter
for each subsequent request, until a limit is reached. The problem here
is that when the fixed window ends, the counter resets to zero, allowing
another burst of requests.
- Sliding window rate limiting by storing each event in the DB with a
timestamp. This means storing each event, either in its own row or in an
array field. Then when a new event occurs, we check to see if the total
events within the window exceeds the limit. It's also necessary to
delete expired events once they're no longer relevant.

Another approach is to maintain a count of allowed actions, and to
decrement the count each time the action is performed. If the count hits
zero, the action is rejected. We can also maintain a timestamp of the
last time the action count was "refilled". So if we are processing a new
action, and the previous action was 30 minutes ago, then the net update
is -1 for the current action being performed and +5 for the "refill"
based on a limit of 10 actions per hour. This allows us to keep an
accurate limit on the number of actions performed per hour without
needing to track each individual action.

This is conceptually neat, but maybe it's a bit harder to grok than it
needs to be.

The API
(https://github.com/proofcarryingdata/zupass/pull/1264/files#diff-124ca6d1a194037b243aa650a70c09b44f0537c8d9671dc10311ee96accfbb3b)
is very simple, and the tests are also very simple
(https://github.com/proofcarryingdata/zupass/pull/1264/files#diff-345b2f6a97b2a26a820f72c77332174ab35843da414d9d4a7d2c74bae36ee21c),
since most of the complexity is handled by a stored procedure. This also
has the advantage of ensuring a single round trip to the DB for each
rate-limiting request.